Divide up the code in two files
