+++ dgraph/cmd/zero/run.go
@@ -256,7 +256,8 @@ func run() {
 	// Create and initialize write-ahead log.
 	x.Checkf(os.MkdirAll(opts.w, 0700), "Error while creating WAL dir.")
 	store := raftwal.Init(opts.w)
-	store.SetRaftId(opts.nodeId)
+	store.SetUint(raftwal.RaftId, opts.nodeId)
+	store.SetUint(raftwal.GroupId, 0) // All zeros have group zero.
 
 	// Initialize the servers.
 	var st state
