+++ raftwal/storage.go
@@ -18,17 +18,12 @@ package raftwal
 
 import (
 	"encoding/binary"
-	"fmt"
 	"math"
 	"os"
 	"path"
 	"path/filepath"
-	"sort"
-	"strconv"
-	"strings"
 	"sync"
 
-	"github.com/dgraph-io/dgraph/protos/pb"
 	"github.com/dgraph-io/dgraph/x"
 	"github.com/dgraph-io/ristretto/z"
 	"github.com/golang/glog"
@@ -97,16 +92,40 @@ type indexRange struct {
 	from, until uint64 // index range for deletion, until index is not deleted.
 }
 
+type MetaInfo int
+
+const (
+	RaftId MetaInfo = iota
+	GroupId
+	CheckpointIndex
+	SnapshotIndex
+	SnapshotTerm
+)
+
+// getOffset returns offsets in wal.meta file.
+func getOffset(info MetaInfo) int {
+	switch info {
+	case RaftId:
+		return 0
+	case GroupId:
+		return 8
+	case CheckpointIndex:
+		return 16
+	case SnapshotIndex:
+		return snapshotIndex
+	case SnapshotTerm:
+		return snapshotIndex + 8
+	default:
+		panic("Invalid info" + string(info))
+	}
+}
+
 // Constants to use when writing to mmap'ed meta and entry files.
 const (
 	// metaName is the name of the file used to store metadata (e.g raft ID, checkpoint).
 	metaName = "wal.meta"
 	// metaFileSize is the size of the wal.meta file.
 	metaFileSize = 4 << 30
-	// raftIdOffset is the offset of the raft ID within the wal.meta file.
-	raftIdOffset = 0
-	// checkpointOffset is the offset of the checkpoint within the wal.meta file.
-	checkpointOffset = 8
 	//hardStateOffset is the offset of the hard sate within the wal.meta file.
 	hardStateOffset = 512
 	// snapshotIndex stores the index and term corresponding to the snapshot.
@@ -115,16 +134,6 @@ const (
 	snapshotOffset = snapshotIndex + 16
 	// maxNumEntries is maximum number of entries before rotating the file.
 	maxNumEntries = 30000
-	// entryFileOffset
-	entryFileOffset = 1 << 20 // 1MB
-	// entryFileSize is the initial size of the entry file.
-	entryFileSize = 16 << 30
-	// entrySize is the size in bytes of a single entry.
-	entrySize = 32
-)
-
-var (
-	emptyEntry = entry(make([]byte, entrySize))
 )
 
 type mmapFile struct {
@@ -254,24 +263,15 @@ func readSlice(dst []byte, offset int) []byte {
 	return b[4 : 4+sz]
 }
 
-func (m *metaFile) raftBuf() []byte {
-	return m.data[raftIdOffset : raftIdOffset+8]
+func (m *metaFile) bufAt(info MetaInfo) []byte {
+	pos := getOffset(info)
+	return m.data[pos : pos+8]
 }
-
-func (m *metaFile) RaftId() uint64 {
-	return binary.BigEndian.Uint64(m.raftBuf())
+func (m *metaFile) Uint(info MetaInfo) uint64 {
+	return binary.BigEndian.Uint64(m.bufAt(info))
 }
-
-func (m *metaFile) StoreRaftId(id uint64) {
-	binary.BigEndian.PutUint64(m.raftBuf(), id)
-}
-
-func (m *metaFile) UpdateCheckpoint(index uint64) {
-	binary.BigEndian.PutUint64(m.data[checkpointOffset:], index)
-}
-
-func (m *metaFile) Checkpoint() uint64 {
-	return binary.BigEndian.Uint64(m.data[checkpointOffset:])
+func (m *metaFile) SetUint(info MetaInfo, id uint64) {
+	binary.BigEndian.PutUint64(m.bufAt(info), id)
 }
 
 func (m *metaFile) StoreHardState(hs *raftpb.HardState) error {
@@ -304,8 +304,8 @@ func (m *metaFile) StoreSnapshot(snap *raftpb.Snapshot) error {
 	if snap == nil || raft.IsEmptySnap(*snap) {
 		return nil
 	}
-	binary.BigEndian.PutUint64(m.data[snapshotIndex:], snap.Metadata.Index)
-	binary.BigEndian.PutUint64(m.data[snapshotIndex+8:], snap.Metadata.Term)
+	m.SetUint(SnapshotIndex, snap.Metadata.Index)
+	m.SetUint(SnapshotTerm, snap.Metadata.Term)
 
 	buf, err := snap.Marshal()
 	if err != nil {
@@ -320,10 +320,6 @@ func (m *metaFile) StoreSnapshot(snap *raftpb.Snapshot) error {
 	return nil
 }
 
-func (m *metaFile) snapshotIndex() uint64 {
-	return binary.BigEndian.Uint64(m.data[snapshotOffset:])
-}
-
 func (m *metaFile) snapshot() (raftpb.Snapshot, error) {
 	val := readSlice(m.data, snapshotOffset)
 
@@ -336,583 +332,6 @@ func (m *metaFile) snapshot() (raftpb.Snapshot, error) {
 		return snap, errors.Wrapf(err, "cannot parse snapshot")
 	}
 	return snap, nil
-}
-
-type entry []byte
-
-func (e entry) Term() uint64 {
-	return binary.BigEndian.Uint64(e)
-}
-func (e entry) Index() uint64 {
-	return binary.BigEndian.Uint64(e[8:])
-}
-func (e entry) DataOffset() uint64 {
-	return binary.BigEndian.Uint64(e[16:])
-}
-func (e entry) Type() uint64 {
-	return binary.BigEndian.Uint64(e[24:])
-}
-
-func marshalEntry(b []byte, term, index, do, typ uint64) {
-	x.AssertTrue(len(b) == entrySize)
-
-	binary.BigEndian.PutUint64(b, term)
-	binary.BigEndian.PutUint64(b[8:], index)
-	binary.BigEndian.PutUint64(b[16:], do)
-	binary.BigEndian.PutUint64(b[24:], typ)
-}
-
-// entryFile represents a single log file.
-type entryFile struct {
-	*mmapFile
-	fid int64
-}
-
-func (ef *entryFile) delete() error {
-	glog.V(2).Infof("Deleting file: %s\n", ef.fd.Name())
-	if err := z.Munmap(ef.data); err != nil {
-		glog.Errorf("while munmap file: %s, error: %v\n", ef.fd.Name(), err)
-	}
-	if err := ef.fd.Truncate(0); err != nil {
-		glog.Errorf("while truncate file: %s, error: %v\n", ef.fd.Name(), err)
-	}
-	return os.Remove(ef.fd.Name())
-}
-
-func openEntryFile(dir string, fid int64) (*entryFile, error) {
-	glog.V(2).Infof("opening entry file: %d\n", fid)
-	fpath := entryFileName(dir, fid)
-	mf, err := openMmapFile(fpath, os.O_RDWR|os.O_CREATE, entryFileSize)
-
-	if err == errNewFile {
-		glog.V(2).Infof("New file: %d\n", fid)
-		zeroOut(mf.data, 0, entryFileOffset)
-	} else {
-		x.Check(err)
-	}
-
-	// } else if err != nil {
-	// 	return nil, err
-	// } else {
-	// }
-	ef := &entryFile{
-		mmapFile: mf,
-		fid:      fid,
-	}
-	return ef, nil
-}
-
-func getEntryFiles(dir string) ([]*entryFile, error) {
-	entryFiles := x.WalkPathFunc(dir, func(path string, isDir bool) bool {
-		if isDir {
-			return false
-		}
-		if strings.HasSuffix(path, ".ent") {
-			return true
-		}
-		return false
-	})
-
-	var files []*entryFile
-	seen := make(map[int64]struct{})
-	for _, fpath := range entryFiles {
-		_, fname := path.Split(fpath)
-		fname = strings.TrimSuffix(fname, ".ent")
-		fid, err := strconv.ParseInt(fname, 10, 64)
-		if err != nil {
-			return nil, errors.Wrapf(err, "while parsing: %s", fpath)
-		}
-		if _, ok := seen[fid]; ok {
-			glog.Fatalf("Entry file with id: %d is repeated", fid)
-		}
-		seen[fid] = struct{}{}
-
-		f, err := openEntryFile(dir, fid)
-		if err != nil {
-			return nil, err
-		}
-		glog.Infof("Found file: %d First Index: %d\n", fid, f.firstIndex())
-		files = append(files, f)
-	}
-
-	// Sort files by the first index they store.
-	sort.Slice(files, func(i, j int) bool {
-		return files[i].firstEntry().Index() < files[j].firstEntry().Index()
-	})
-	return files, nil
-}
-
-// get entry from a file.
-func (ef *entryFile) getEntry(idx int) entry {
-	if ef == nil {
-		return emptyEntry
-	}
-	offset := idx * entrySize
-	return entry(ef.data[offset : offset+entrySize])
-}
-
-func (ef *entryFile) GetRaftEntry(idx int) raftpb.Entry {
-	entry := ef.getEntry(idx)
-	re := raftpb.Entry{
-		Term:  entry.Term(),
-		Index: entry.Index(),
-		Type:  raftpb.EntryType(int32(entry.Type())),
-	}
-	if entry.DataOffset() > 0 && entry.DataOffset() < entryFileSize {
-		data := ef.slice(int(entry.DataOffset()))
-		if len(data) > 0 {
-			re.Data = append(re.Data, data...)
-		}
-	}
-	return re
-}
-
-func (ef *entryFile) firstEntry() entry {
-	return ef.getEntry(0)
-}
-func (ef *entryFile) firstIndex() uint64 {
-	return ef.getEntry(0).Index()
-}
-func (ef *entryFile) firstEmptySlot() int {
-	return sort.Search(maxNumEntries, func(i int) bool {
-		e := ef.getEntry(i)
-		return e.Index() == 0
-	})
-}
-func (ef *entryFile) lastEntry() entry {
-	// This would return the first pos, where e.Index() == 0.
-	pos := ef.firstEmptySlot()
-	if pos > 0 {
-		pos--
-	}
-	return ef.getEntry(pos)
-}
-
-func (ef *entryFile) Term(entryIndex uint64) uint64 {
-	offset := ef.slotGe(entryIndex)
-	if offset < 0 || offset >= maxNumEntries {
-		return 0
-	}
-	e := ef.getEntry(int(offset))
-	if e.Index() == entryIndex {
-		return e.Term()
-	}
-	return 0
-}
-
-// slotGe would return -1 if raftIndex < firstIndex in this file.
-// Would return maxNumEntries if raftIndex > lastIndex in this file.
-// If raftIndex is found, or the entryFile has empty slots, the offset would be between
-// [0, maxNumEntries).
-func (ef *entryFile) slotGe(raftIndex uint64) int {
-	fi := ef.firstIndex()
-	// If first index is zero, this raftindex should be in a previous file.
-	if fi == 0 {
-		return -1
-	}
-	if raftIndex < fi {
-		return -1
-	}
-	if diff := int(raftIndex - fi); diff < maxNumEntries && diff >= 0 {
-		e := ef.getEntry(diff)
-		if e.Index() == raftIndex {
-			return diff
-		}
-	}
-
-	// This would find the first entry's index which has entryIndex.
-	return sort.Search(maxNumEntries, func(i int) bool {
-		e := ef.getEntry(i)
-		if e.Index() == 0 {
-			// We reached too far to the right.
-			return true
-		}
-		return e.Index() >= raftIndex
-	})
-}
-
-// entryLog represents the entire entry log. It consists of one or more
-// entryFile objects.
-type entryLog struct {
-	// need lock for files and current ?
-
-	// files is the list of all log files ordered in ascending order by the first
-	// index in the file. The current file being written should always be accessible
-	// by looking at the last element of this slice.
-	files   []*entryFile
-	current *entryFile
-	// nextEntryIdx is the index of the next entry to write to. When this value exceeds
-	// maxNumEntries the file will be rotated.
-	nextEntryIdx int
-	// dir is the directory to use to store files.
-	dir string
-}
-
-func entryFileName(dir string, id int64) string {
-	return path.Join(dir, fmt.Sprintf("%05d.ent", id))
-}
-
-func openEntryLog(dir string) (*entryLog, error) {
-	e := &entryLog{
-		dir: dir,
-	}
-	files, err := getEntryFiles(dir)
-	if err != nil {
-		return nil, err
-	}
-	out := files[:0]
-	var nextFid int64
-	for _, ef := range files {
-		if nextFid < ef.fid {
-			nextFid = ef.fid
-		}
-		if ef.firstIndex() == 0 {
-			if err := ef.delete(); err != nil {
-				return nil, err
-			}
-		} else {
-			out = append(out, ef)
-		}
-	}
-	e.files = out
-	if sz := len(e.files); sz > 0 {
-		e.current = e.files[sz-1]
-		e.nextEntryIdx = e.current.firstEmptySlot()
-
-		e.files = e.files[:sz-1]
-		return e, nil
-	}
-
-	// No files found. Create a new file.
-	nextFid += 1
-	ef, err := openEntryFile(dir, nextFid)
-	e.current = ef
-	return e, err
-}
-
-func (l *entryLog) lastFile() *entryFile {
-	return l.files[len(l.files)-1]
-}
-
-// getEntry gets the nth entry in the CURRENT log file.
-func (l *entryLog) getEntry(n int) (entry, error) {
-	if n >= maxNumEntries {
-		return nil, errors.Errorf("there cannot be more than %d in a single file",
-			maxNumEntries)
-	}
-
-	start := n * entrySize
-	buf := l.current.data[start : start+entrySize]
-	return entry(buf), nil
-}
-
-func (l *entryLog) rotate(firstIndex uint64) error {
-	nextFid := l.current.fid
-	x.AssertTrue(nextFid > 0)
-
-	for _, ef := range l.files {
-		if ef.fid > nextFid {
-			nextFid = ef.fid
-		}
-	}
-
-	nextFid += 1
-	ef, err := openEntryFile(l.dir, nextFid)
-	if err != nil {
-		return errors.Wrapf(err, "while creating a new entry file")
-	}
-
-	l.files = append(l.files, l.current)
-	l.current = ef
-	return nil
-}
-
-func (l *entryLog) numEntries() int {
-	if len(l.files) == 0 {
-		return 0
-	}
-	total := 0
-	if len(l.files) >= 1 {
-		// all files except the last one.
-		total += (len(l.files) - 1) * maxNumEntries
-	}
-	return total + l.nextEntryIdx
-}
-
-func (l *entryLog) AddEntries(entries []raftpb.Entry) error {
-	if len(entries) == 0 {
-		return nil
-	}
-	// glog.Infof("AddEntries: %+v\n", entries)
-	fidx, eidx := l.slotGe(entries[0].Index)
-
-	// fmt.Printf("fidx: %d, eidx: %d, entries: %+v\n", fidx, eidx, entries)
-	if eidx >= 0 {
-		// fmt.Printf("AddEntries: fidx: %d, eidx: %d num: %d\n", fidx, eidx, len(entries))
-
-		if fidx == -1 {
-			if l.nextEntryIdx > eidx {
-				zeroOut(l.current.data, entrySize*eidx, entrySize*l.nextEntryIdx)
-			}
-		} else {
-			x.AssertTrue(fidx < len(l.files))
-			extra := l.files[fidx+1:]
-			extra = append(extra, l.current)
-			l.current = l.files[fidx]
-
-			for _, ef := range extra {
-				glog.V(2).Infof("Deleting extra file: %d\n", ef.fid)
-				if err := ef.delete(); err != nil {
-					glog.Errorf("deleting file: %s. error: %v\n", ef.fd.Name(), err)
-				}
-			}
-			zeroOut(l.current.data, entrySize*eidx, entryFileOffset)
-			l.files = l.files[:fidx]
-		}
-		l.nextEntryIdx = eidx
-	}
-
-	prev := l.nextEntryIdx - 1
-	var offset int
-	if prev >= 0 {
-		e := l.current.getEntry(prev)
-		offset = int(e.DataOffset())
-		offset += sliceSize(l.current.data, offset)
-	} else {
-		offset = entryFileOffset
-	}
-
-	for _, re := range entries {
-		if l.nextEntryIdx >= maxNumEntries {
-			if err := l.rotate(re.Index); err != nil {
-				// TODO: see what happens.
-				return err
-			}
-			l.nextEntryIdx, offset = 0, entryFileOffset
-		}
-
-		destBuf, next := l.current.allocateSlice(len(re.Data), offset)
-		x.AssertTrue(copy(destBuf, re.Data) == len(re.Data))
-
-		buf, err := l.getEntry(l.nextEntryIdx)
-		x.Check(err)
-		marshalEntry(buf, re.Term, re.Index, uint64(offset), uint64(re.Type))
-
-		// Update for next entry.
-		offset = next
-		l.nextEntryIdx++
-	}
-	return nil
-}
-
-func (l *entryLog) DiscardFiles(snapshotIndex uint64) error {
-	// TODO: delete all the files below the first file with a first index
-	// less than or equal to snapshotIndex.
-	return nil
-}
-
-func (l *entryLog) firstIndex() uint64 {
-	if l == nil {
-		return 0
-	}
-	var fi uint64
-	if len(l.files) == 0 {
-		fi = l.current.firstEntry().Index()
-	} else {
-		fi = l.files[0].firstEntry().Index()
-	}
-	if fi == 0 {
-		return 1
-	}
-	return fi
-}
-
-func (l *entryLog) LastIndex() uint64 {
-	if l.nextEntryIdx-1 >= 0 {
-		e := l.current.getEntry(l.nextEntryIdx - 1)
-		return e.Index()
-	}
-	for i := len(l.files) - 1; i >= 0; i-- {
-		ef := l.files[i]
-		e := ef.lastEntry()
-		if e.Index() > 0 {
-			return e.Index()
-		}
-	}
-	return 0
-}
-
-func (l *entryLog) getEntryFile(fidx int) *entryFile {
-	if fidx == -1 {
-		return l.current
-	}
-	if fidx >= len(l.files) {
-		return nil
-	}
-	return l.files[fidx]
-}
-
-func (l *entryLog) seekEntry(raftIndex uint64) (entry, error) {
-	if raftIndex == 0 {
-		return emptyEntry, nil
-	}
-
-	fidx, off := l.slotGe(raftIndex)
-	if off == -1 {
-		return emptyEntry, raft.ErrCompacted
-	} else if off >= maxNumEntries {
-		return emptyEntry, raft.ErrUnavailable
-	}
-	// idx := ef.getEntry(off).Index()
-
-	// if off == -1 {
-	// 	if raftIndex < ef.firstEntry().Index() {
-	// 		return emptyEntry, raft.ErrCompacted
-	// 	}
-	// 	return emptyEntry, raft.ErrUnavailable
-	// }
-
-	ef := l.getEntryFile(fidx)
-	ent := ef.getEntry(off)
-	if ent.Index() == 0 {
-		// We have gone past what we wrote to the file.
-		return emptyEntry, raft.ErrUnavailable
-	}
-	if ent.Index() != raftIndex {
-		return emptyEntry, errNotFound
-	}
-	return ent, nil
-}
-
-// Term returns the term of entry i, which must be in the range
-// [FirstIndex()-1, LastIndex()]. The term of the entry before
-// FirstIndex is retained for matching purposes even though the
-// rest of that entry may not be available.
-func (l *entryLog) Term(idx uint64) (uint64, error) {
-	// Look at the entry files and find the entry file with entry bigger than idx.
-	// Read file before that idx.
-	ent, err := l.seekEntry(idx)
-	return ent.Term(), err
-}
-
-// slotGe returns the file index and the slot within that file in which the entry
-// with the given index can be found. A value of -1 for the file index means that the
-// entry is in the current file.
-func (l *entryLog) slotGe(raftIndex uint64) (int, int) {
-	// Look for the offset in the current file.
-	if offset := l.current.slotGe(raftIndex); offset >= 0 {
-		return -1, offset
-	}
-
-	// No previous files, therefore we can only go back to the start of the current file.
-	if len(l.files) == 0 {
-		return -1, -1
-	}
-
-	fileIdx := sort.Search(len(l.files), func(i int) bool {
-		return l.files[i].firstIndex() >= raftIndex
-	})
-	if fileIdx >= len(l.files) {
-		fileIdx = len(l.files) - 1
-	}
-	for fileIdx > 0 {
-		fi := l.files[fileIdx].firstIndex()
-		if fi <= raftIndex {
-			break
-		}
-		fileIdx--
-	}
-	offset := l.files[fileIdx].slotGe(raftIndex)
-	return fileIdx, offset
-}
-
-func (l *entryLog) deleteBefore(raftIndex uint64) error {
-	fidx, off := l.slotGe(raftIndex)
-
-	if off >= 0 && fidx <= len(l.files) {
-		var before []*entryFile
-		if fidx == -1 { // current file
-			before = l.files
-			l.files = l.files[:0]
-		} else {
-			before = l.files[:fidx]
-			l.files = l.files[fidx:]
-		}
-
-		for _, ef := range before {
-			if err := ef.delete(); err != nil {
-				glog.Errorf("while deleting file: %s, err: %v\n", ef.fd.Name(), err)
-			}
-		}
-	}
-	return nil
-}
-
-func (l *entryLog) reset() error {
-	for _, ef := range l.files {
-		if err := ef.delete(); err != nil {
-			return errors.Wrapf(err, "while deleting %s", ef.fd.Name())
-		}
-	}
-	l.files = l.files[:0]
-	zeroOut(l.current.data, 0, entryFileOffset)
-	var num int
-	for _, b := range l.current.data[:entryFileOffset] {
-		x.AssertTrue(b == 0x00)
-		num++
-	}
-	l.nextEntryIdx = 0
-	return nil
-}
-
-func (l *entryLog) allEntries(lo, hi, maxSize uint64) []raftpb.Entry {
-	var entries []raftpb.Entry
-	fileIdx, offset := l.slotGe(lo)
-	var size uint64
-
-	if offset < 0 {
-		// We are at the very beginning of this thing.
-		offset = 0
-	}
-
-	currFile := l.getEntryFile(fileIdx)
-	for {
-		if offset >= maxNumEntries {
-			if fileIdx == -1 {
-				// We are looking at the current file and there are no more entries.
-				// Return what we have.
-				return entries
-			}
-
-			// Move to the next file.
-			fileIdx++
-			if fileIdx >= len(l.files) {
-				fileIdx = -1
-			}
-			currFile = l.getEntryFile(fileIdx)
-			x.AssertTrue(currFile != nil)
-
-			// Reset the offset to start reading the next file from the beginning.
-			offset = 0
-		}
-
-		re := currFile.GetRaftEntry(offset)
-		if re.Index >= hi {
-			return entries
-		}
-		if re.Index == 0 {
-			// Allow this to move to the next file.
-			offset = maxNumEntries
-			continue
-		}
-		size += uint64(re.Size())
-		if len(entries) > 0 && size > maxSize {
-			break
-		}
-		entries = append(entries, re)
-		offset++
-	}
-	return entries
 }
 
 // Init initializes returns a properly initialized instance of DiskStorage.
@@ -935,9 +354,12 @@ func Init(dir string) *DiskStorage {
 
 	snap, err := w.meta.snapshot()
 	x.Check(err)
-	glog.Infof("Found snapshot: %+v\n", snap)
 
 	first, _ := w.FirstIndex()
+	if !raft.IsEmptySnap(snap) {
+		x.AssertTruef(snap.Metadata.Index+1 == first,
+			"snap index: %d + 1 should be equal to first: %d\n", snap.Metadata.Index, first)
+	}
 
 	// If db is not closed properly, there might be index ranges for which delete entries are not
 	// inserted. So insert delete entries for those ranges starting from 0 to (first-1).
@@ -951,31 +373,14 @@ func Init(dir string) *DiskStorage {
 	return w
 }
 
-// TODO: Have a way to set and request gid.
-func (w *DiskStorage) SetRaftId(raftId uint64) {
-	w.meta.StoreRaftId(raftId)
+func (w *DiskStorage) SetUint(info MetaInfo, id uint64) {
+	w.meta.SetUint(info, id)
 }
-func (w *DiskStorage) RaftId() uint64 {
-	return w.meta.RaftId()
+func (w *DiskStorage) Uint(info MetaInfo) uint64 {
+	return w.meta.Uint(info)
 }
 
 var errNotFound = errors.New("Unable to find raft entry")
-
-// setSnapshot would store the snapshot. We can delete all the entries up until the snapshot
-// index. But, keep the raft entry at the snapshot index, to make it easier to build the logic; like
-// the dummy entry in MemoryStorage.
-// func (w *DiskStorage) setSnapshot(s *raftpb.Snapshot) error {
-// 	if s == nil || raft.IsEmptySnap(*s) {
-// 		return nil
-// 	}
-
-// 	if err := w.meta.StoreSnapshot(s); err != nil {
-// 		return err
-// 	}
-// 	// TODO: Do we need to overwrite the entry?
-// 	// e := raftpb.Entry{Term: s.Metadata.Term, Index: s.Metadata.Index}
-// 	return nil
-// }
 
 // reset resets the entries. Used for testing.
 func (w *DiskStorage) reset(es []raftpb.Entry) error {
@@ -992,20 +397,11 @@ func (w *DiskStorage) HardState() (raftpb.HardState, error) {
 	}
 	return w.meta.HardState()
 }
-
 func (w *DiskStorage) Checkpoint() (uint64, error) {
 	if w.meta == nil {
 		return 0, errors.Errorf("uninitialized meta file")
 	}
-	return w.meta.Checkpoint(), nil
-}
-
-func (w *DiskStorage) UpdateCheckpoint(snap *pb.Snapshot) error {
-	if w.meta == nil {
-		return errors.Errorf("uninitialized meta file")
-	}
-	w.meta.UpdateCheckpoint(snap.Index)
-	return nil
+	return w.meta.Uint(CheckpointIndex), nil
 }
 
 // Implement the Raft.Storage interface.
@@ -1105,7 +501,7 @@ func (w *DiskStorage) firstIndex() uint64 {
 	// We are deleting index ranges in background after taking snapshot, so we should check for last
 	// snapshot in WAL(Badger) if it is not found in cache. If no snapshot is found, then we can
 	// check firstKey.
-	if si := w.meta.snapshotIndex(); si > 0 {
+	if si := w.Uint(SnapshotIndex); si > 0 {
 		return si + 1
 	}
 	return w.entries.firstIndex()
